Modularized, error fixes
