+++ PyEDF-APP/.gitignore
@@ -1,0 +1,3 @@
+*.pyc
+__pycache__
+
